Move `utils.js` to `utils/index.js`
